test: split supervisor coverage by responsibility
Split the 453-line supervisor test module into three focused modules so every module remains below the repository hard 200-line limit. Preserve all ten test ASTs and leave production behavior unchanged.

Verification: make ci-pr (Ruff, strict MyPy, 648 unit tests with 1 skip, and 25 sealed golden tests).

Human verification: VrtxOmega
